wip : handle actions

+++ src/client/interfaces/game_ui/game_ui.h
@@ -5,7 +5,7 @@
 #include <gtk/gtk.h>
 
 void untoggle_previous_bet_btn(GtkBuilder *builder,GtkWidget *button);
-void activate_countdown(GtkProgressBar *progress_bar, int seconds, int fps);
+void activate_countdown(GtkProgressBar *progress_bar, double seconds, double fps);
 void toggle_action_button(GtkBuilder *builder, gboolean sensitive);
 void radio_bet_button(GtkBuilder *builder, gboolean sensitive);
 void stop_count_down(GtkProgressBar *progress_bar);
